refactor: consolidate kestraClient() and kestraUrl into AbstractCloningTask
Move the kestraClient() factory method and kestraUrl property from each
individual task class into AbstractCloningTask so that all tasks inherit
them without duplication. This also enables tests to override the Kestra
API URL via .kestraUrl(Property.ofValue(mockServerUrl)).
