+++ src/main/java/io/kestra/plugin/git/AbstractCloningTask.java
@@ -2,6 +2,7 @@ package io.kestra.plugin.git;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.lib.ObjectId;
@@ -13,7 +14,11 @@ import org.eclipse.jgit.transport.RefSpec;
 import org.eclipse.jgit.transport.TagOpt;
 import org.slf4j.Logger;
 
+import io.kestra.core.exceptions.IllegalVariableEvaluationException;
 import io.kestra.core.models.property.Property;
+import io.kestra.core.runners.RunContext;
+import io.kestra.core.runners.SDK;
+import io.kestra.sdk.KestraClient;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
@@ -25,12 +30,51 @@ import io.kestra.core.models.annotations.PluginProperty;
 @NoArgsConstructor
 @Getter
 public abstract class AbstractCloningTask extends AbstractGitTask {
+    private static final String DEFAULT_KESTRA_URL = "http://localhost:8080";
+    private static final String KESTRA_URL_TEMPLATE = "{{ kestra.url }}";
+
+    @Schema(
+        title = "Kestra API URL",
+        description = """
+            URL of the Kestra server API.
+            If not set, the value of `kestra.url` from the Kestra configuration is used.
+            If that is also not set, defaults to `http://localhost:8080`."""
+    )
+    @PluginProperty(group = "connection")
+    protected Property<String> kestraUrl;
+
     @Schema(
         title = "Clone submodules",
         description = "Default false; enable to fetch and initialize nested submodules."
     )
     @PluginProperty(group = "advanced")
     protected Property<Boolean> cloneSubmodules;
+
+    protected KestraClient kestraClient(RunContext runContext) throws IllegalVariableEvaluationException {
+        String rKestraUrl = runContext.render(kestraUrl).as(String.class)
+            .orElseGet(() -> {
+                try {
+                    String rendered = runContext.render(KESTRA_URL_TEMPLATE);
+                    return (rendered == null || rendered.isBlank()) ? DEFAULT_KESTRA_URL : rendered;
+                } catch (IllegalVariableEvaluationException e) {
+                    return DEFAULT_KESTRA_URL;
+                }
+            });
+
+        if (rKestraUrl == null || rKestraUrl.isBlank()) {
+            rKestraUrl = DEFAULT_KESTRA_URL;
+        }
+
+        runContext.logger().debug("Kestra URL: {}", rKestraUrl);
+
+        var normalizedUrl = rKestraUrl.trim().replaceAll("/+$", "");
+        var builder = KestraClient.builder().url(normalizedUrl);
+        Optional<SDK.Auth> autoAuth = runContext.sdk().defaultAuthentication();
+        if (autoAuth.isPresent() && autoAuth.get().username().isPresent() && autoAuth.get().password().isPresent()) {
+            return builder.basicAuth(autoAuth.get().username().get(), autoAuth.get().password().get()).build();
+        }
+        return builder.build();
+    }
 
     protected void checkoutCommit(Git git, String sha, Logger logger, boolean noTags) throws Exception {
         // Ensure we have a full history in case the repo was shallow by default on the remote
