+++ src/main/java/io/kestra/plugin/git/PushFlows.java
@@ -25,7 +25,6 @@ import io.kestra.core.models.annotations.PluginProperty;
 import io.kestra.core.models.flows.FlowWithSource;
 import io.kestra.core.models.property.Property;
 import io.kestra.core.runners.RunContext;
-import io.kestra.core.runners.SDK;
 import io.kestra.core.serializers.YamlParser;
 import io.kestra.sdk.KestraClient;
 import io.kestra.sdk.internal.ApiException;
@@ -257,25 +256,6 @@ public class PushFlows extends AbstractPushTask<PushFlows.Output> {
         } catch (ApiException e) {
             throw new KestraRuntimeException("Failed to export flows from Kestra for namespace " + namespace, e);
         }
-    }
-
-    private KestraClient kestraClient(RunContext runContext) throws IllegalVariableEvaluationException {
-        String rKestraUrl;
-        try {
-            rKestraUrl = runContext.render("{{ kestra.url }}");
-        } catch (IllegalVariableEvaluationException e) {
-            rKestraUrl = "http://localhost:8080";
-        }
-        if (rKestraUrl == null || rKestraUrl.isBlank()) {
-            rKestraUrl = "http://localhost:8080";
-        }
-        String normalizedUrl = rKestraUrl.trim().replaceAll("/+$", "");
-        var builder = KestraClient.builder().url(normalizedUrl);
-        Optional<SDK.Auth> autoAuth = runContext.sdk().defaultAuthentication();
-        if (autoAuth.isPresent() && autoAuth.get().username().isPresent() && autoAuth.get().password().isPresent()) {
-            return builder.basicAuth(autoAuth.get().username().get(), autoAuth.get().password().get()).build();
-        }
-        return builder.build();
     }
 
     @Override
