+++ src/main/java/io/kestra/plugin/git/SyncFlows.java
@@ -26,7 +26,6 @@ import io.kestra.core.models.flows.FlowWithException;
 import io.kestra.core.models.flows.FlowWithSource;
 import io.kestra.core.models.property.Property;
 import io.kestra.core.runners.RunContext;
-import io.kestra.core.runners.SDK;
 import io.kestra.core.serializers.YamlParser;
 import io.kestra.sdk.KestraClient;
 import io.kestra.sdk.internal.ApiException;
@@ -409,25 +408,6 @@ public class SyncFlows extends AbstractSyncTask<Flow, SyncFlows.Output> {
         } catch (IOException e) {
             throw new KestraRuntimeException("Failed to create named file for: " + fileName, e);
         }
-    }
-
-    private KestraClient kestraClient(RunContext runContext) throws IllegalVariableEvaluationException {
-        String rKestraUrl;
-        try {
-            rKestraUrl = runContext.render("{{ kestra.url }}");
-        } catch (IllegalVariableEvaluationException e) {
-            rKestraUrl = "http://localhost:8080";
-        }
-        if (rKestraUrl == null || rKestraUrl.isBlank()) {
-            rKestraUrl = "http://localhost:8080";
-        }
-        String normalizedUrl = rKestraUrl.trim().replaceAll("/+$", "");
-        var builder = KestraClient.builder().url(normalizedUrl);
-        Optional<SDK.Auth> autoAuth = runContext.sdk().defaultAuthentication();
-        if (autoAuth.isPresent() && autoAuth.get().username().isPresent() && autoAuth.get().password().isPresent()) {
-            return builder.basicAuth(autoAuth.get().username().get(), autoAuth.get().password().get()).build();
-        }
-        return builder.build();
     }
 
     @Override
